CipherTalkV2/CipherTalkV2/requirements.txt
pynacl
cryptography
sqlalchemy
typer
python-dotenv
